corrected seekpos cast bug

+++ streamer.c
@@ -1551,7 +1551,7 @@ streamer_thread (void *ctx) {
             streamer_unlock ();
         }
 
-        int seek = seekpos;
+        float seek = seekpos;
         seekpos = -1;
         if (seek >= 0 && pl_get_item_duration (playing_track) > 0) {
             playpos = seek;
